+++ packages/ts-newskit/src/slices/content-bucket-1/styles.ts
@@ -1,7 +1,0 @@
-import { Divider, styled } from 'newskit';
-
-export const ArticleDividerXL = styled(Divider)`
-  position: absolute;
-  height: 100%;
-  left: 50%;
-`;