+++ packages/ts-newskit/src/slices/content-bucket-2/__tests__/index.test.tsx
@@ -37,6 +37,12 @@ describe('Render Content Bucket 2 Slice', () => {
     const { asFragment } = await renderComponent();
     expect(asFragment()).toMatchSnapshot();
   });
+
+  test('Slice matches snapshot for `null` breakpoint value', async () => {
+    (useBreakpointKey as any).mockReturnValue(null);
+    const { asFragment } = await renderComponent();
+    expect(asFragment()).toMatchSnapshot();
+  });
 });
 
 describe('Content Bucket 2 Articles list above `md` breakpoint', () => {