+++ packages/ts-newskit/src/slices/lead-story-2/__tests__/index.test.tsx
@@ -42,4 +42,9 @@ describe('Render Lead Story 2 Slice', () => {
     const { asFragment } = renderComponent();
     expect(asFragment()).toMatchSnapshot();
   });
+  test('Slice matches snapshot for `null` breakpoint value', () => {
+    (useBreakpointKey as any).mockReturnValue(null);
+    const { asFragment } = renderComponent();
+    expect(asFragment()).toMatchSnapshot();
+  });
 });