+++ packages/ts-newskit/src/slices/shared/layouts.tsx
@@ -11,6 +11,7 @@ export const CustomStackLayout: React.FC = ({ children }) => {
         xs: 'space045',
         md: 'space000'
       }}
+      marginBlockEnd="space060"
     >
       {children}
     </Stack>