+++ packages/ts-newskit/src/slices/stacked-module-1/index.tsx
@@ -20,15 +20,6 @@ type ArticleStackProps = {
   clickHandler: ClickHandlerType;
 };
 
-const fullWidthDivider = (marginBlockStart?: string) => (
-  <Divider
-    overrides={{
-      stylePreset: 'dashedDivider',
-      marginBlockStart: marginBlockStart || 'space000'
-    }}
-  />
-);
-
 const articleStack = ({
   articles,
   marginBlockStart,
@@ -39,7 +30,12 @@ const articleStack = ({
   <CustomStackLayout>
     <StackItem>
       <FullWidthBlock marginBlockEnd="space040">
-        {fullWidthDivider(marginBlockStart)}
+        <Divider
+          overrides={{
+            stylePreset: 'dashedDivider',
+            marginBlockStart: marginBlockStart || 'space000'
+          }}
+        />
       </FullWidthBlock>
     </StackItem>
     <StackItem>
@@ -90,7 +86,9 @@ const articleStack = ({
 );
 
 export const StackModule1 = ({ articles, clickHandler }: StackModule1Props) => {
-  const [currentBreakpoint, setBreakpoint] = useState<BreakpointKeys>('xs');
+  const [currentBreakpoint, setBreakpoint] = useState<BreakpointKeys | null>(
+    null
+  );
   const breakpointKey = useBreakpointKey();
   useEffect(
     () => {
@@ -98,6 +96,11 @@ export const StackModule1 = ({ articles, clickHandler }: StackModule1Props) => {
     },
     [breakpointKey]
   );
+
+  if (!currentBreakpoint) {
+    return null;
+  }
+
   const isMob = currentBreakpoint === 'xs' || currentBreakpoint === 'sm';
   const articlesTop = articles.slice(0, 4);
   const articlesBottom = articles.slice(4);
@@ -119,7 +122,6 @@ export const StackModule1 = ({ articles, clickHandler }: StackModule1Props) => {
         isMob,
         clickHandler
       })}
-      {fullWidthDivider('space040')}
     </BlockItem>
   );
 };